requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
ollama
google-genai
python-dotenv
numpy
bcrypt>=4.1
pyjwt>=2.8
